requirements.txt
Django==1.4.2
Pygments==1.5
argparse==1.2.1
djangorestframework==2.1.6
wsgiref==0.1.2
requests==0.14.2
